+++ ExcelTableConverter.JiraTableConverter/EmphTextStylers/BoldTextStyler.cs
@@ -1,12 +1,12 @@
 ﻿using ExcelTableConverter.TableConverter.EmphTextStylers;
 
-namespace ExcelTableConverter.LatexTableConverter.EmphTextStylers
+namespace ExcelTableConverter.JiraTableConverter.EmphTextStylers
 {
   public class BoldTextStyler : EmphTextStyler
   {
     public override string Style(string text)
     {
-      return string.Format("\\textbf{{{0}}}", text);
+      return string.Format("*{0}*", text);
     }
   }
 }