+++ ExcelTableConverter.JiraTableConverter/EmphTextStylers/ItalicTextStyler.cs
@@ -1,12 +1,12 @@
 ﻿using ExcelTableConverter.TableConverter.EmphTextStylers;
 
-namespace ExcelTableConverter.LatexTableConverter.EmphTextStylers
+namespace ExcelTableConverter.JiraTableConverter.EmphTextStylers
 {
   public class ItalicTextStyler : EmphTextStyler
   {
     public override string Style(string text)
     {
-      return string.Format("\\textit{{{0}}}", text);
+      return string.Format("_{0}_", text);
     }
   }
 }