+++ ExcelTableConverter.AddIn/Properties/Resources.Designer.cs
@@ -81,7 +81,7 @@ namespace ExcelTableConverter.AddIn.Properties {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to converting completed.
+        ///   Looks up a localized string similar to Converting completed..
         /// </summary>
         internal static string ConvertingCompleted {
             get {
@@ -90,7 +90,7 @@ namespace ExcelTableConverter.AddIn.Properties {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Current converter not set.
+        ///   Looks up a localized string similar to Current converter not set..
         /// </summary>
         internal static string ConvertNotSet {
             get {
@@ -114,15 +114,6 @@ namespace ExcelTableConverter.AddIn.Properties {
             get {
                 object obj = ResourceManager.GetObject("Icon", resourceCulture);
                 return ((System.Drawing.Icon)(obj));
-            }
-        }
-        
-        /// <summary>
-        ///   Looks up a localized string similar to specifiy file first.
-        /// </summary>
-        internal static string SpecifyFileFirst {
-            get {
-                return ResourceManager.GetString("SpecifyFileFirst", resourceCulture);
             }
         }
     }